requirements.txt
# Backend Dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
motor>=3.3.0
pymongo>=4.6.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
email-validator>=2.1.0
python-dotenv>=1.0.0
httpx>=0.25.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Optional: Development tools
black>=23.0.0
isort>=5.12.0
